+++ packages/x-adapter-platform/src/schemas/models/result.schema.ts
@@ -23,7 +23,6 @@ export const resultSchema = createMutableSchema<PlatformResult, Result>({
     value: '__prices.current.value',
     originalValue: ({ __prices: rawPrices }) =>
       rawPrices.previous?.value ?? rawPrices.current.value,
-    futureValue: ({ __prices: rawPrices }) => rawPrices.future?.value ?? rawPrices.current.value,
     hasDiscount: ({ __prices: rawPrices }) =>
       rawPrices.current.value < (rawPrices.previous?.value ?? rawPrices.current.value),
   },