+++ packages/x-design-system/CHANGELOG.md
@@ -3,6 +3,13 @@
 All notable changes to this project will be documented in this file.
 See [Conventional Commits](https://conventionalcommits.org) for commit guidelines.
 
+## [0.1.0-alpha.0](https://github.com/empathyco/x/compare/@empathyco/x-design-system@0.0.2-alpha.22...@empathyco/x-design-system@0.1.0-alpha.0) (2026-03-24)
+
+### Features
+
+* **xds:** `layouts` migration (#2057)
+
+
 ## [0.0.2-alpha.22](https://github.com/empathyco/x/compare/@empathyco/x-design-system@0.0.2-alpha.21...@empathyco/x-design-system@0.0.2-alpha.22) (2026-03-17)
 
 **Note:** Version bump only for package @empathyco/x-design-system